replace utils

+++ src/main/java/com/github/ked4ma/atcoder/abc202_na/D_x.kt
@@ -1,14 +1,14 @@
 package com.github.ked4ma.atcoder.abc202_na
 
-import com.github.ked4ma.atcoder.utils.combination.*
 import com.github.ked4ma.atcoder.utils.input.default.*
+import com.github.ked4ma.atcoder.utils.math.combimation.*
 
 fun main() {
     var (A, B, K) = nextLongList()
         .let { (A, B, K) -> Triple(A.toInt(), B.toInt(), K) }
     val sb = StringBuilder()
     while (A > 0 && B > 0) {
-        val k = combination(A - 1 + B, B)
+        val k = combination((A - 1 + B).toLong(), B.toLong())
         if (K <= k) {
             sb.append('a')
             A -= 1
