+++ src/main/java/com/github/ked4ma/atcoder/abc323/B.kt
@@ -1,13 +1,13 @@
 package com.github.ked4ma.atcoder.abc323
 
 import com.github.ked4ma.atcoder.utils.input.default.*
-import com.github.ked4ma.atcoder.utils.input.default.deprecated.*
-import com.github.ked4ma.atcoder.utils.loop.deprecated.*
+import com.github.ked4ma.atcoder.utils.input.deprecated.*
+import com.github.ked4ma.atcoder.utils.repeat.*
 
 // make run <TASK: A/B/...> [BRANCH=feature/<CONTEST: abc000>]
 fun main() {
     val N = nextInt()
-    times(N) { i ->
+    times(N.toLong()) { i ->
         next().let {
             i + 1 to it.filter { c -> c == 'o' }.length
         }
