+++ src/main/java/com/github/ked4ma/atcoder/abc323/C.kt
@@ -1,14 +1,14 @@
 package com.github.ked4ma.atcoder.abc323
 
 import com.github.ked4ma.atcoder.utils.input.default.*
-import com.github.ked4ma.atcoder.utils.input.default.deprecated.*
-import com.github.ked4ma.atcoder.utils.loop.deprecated.*
+import com.github.ked4ma.atcoder.utils.input.deprecated.*
+import com.github.ked4ma.atcoder.utils.repeat.*
 
 // make run <TASK: A/B/...> [BRANCH=feature/<CONTEST: abc000>]
 fun main() {
     val (N, M) = nextIntList()
     val A = nextIntList()
-    val S = times(N) {
+    val S = times(N.toLong()) {
         next()
     }
