+++ src/main/java/com/github/ked4ma/atcoder/abc330/D.kt
@@ -2,13 +2,13 @@ package com.github.ked4ma.atcoder.abc330
 
 import com.github.ked4ma.atcoder.utils.array.deprecated.*
 import com.github.ked4ma.atcoder.utils.input.default.*
-import com.github.ked4ma.atcoder.utils.input.default.deprecated.*
-import com.github.ked4ma.atcoder.utils.loop.deprecated.*
+import com.github.ked4ma.atcoder.utils.input.deprecated.*
+import com.github.ked4ma.atcoder.utils.repeat.*
 
 // make run <TASK: A/B/...> [BRANCH=feature/<CONTEST: abc000>]
 fun main() {
     val N = nextInt()
-    val S = times(N) {
+    val S = times(N.toLong()) {
         next().toCharArray()
     }.toTypedArray()
